Remove mobile-compatible project structure
