+++ crates/tauri-app/Cargo.toml
@@ -47,7 +47,3 @@ tauri-plugin-store = "2.0.0-beta"
 # this feature is used for production builds or when `devPath` points to the filesystem
 # DO NOT REMOVE!!
 custom-protocol = ["tauri/custom-protocol"]
-
-[lib]
-name = "app_lib"
-crate-type = ["staticlib", "cdylib", "rlib"]
